requirements.txt
discord.py==2.5.2
folium==0.19.5
geopy==2.4.1
pycountry_convert==0.7.2
python-dotenv==1.0.1
